pyproject.toml
[project]
name = "agent-pikm"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "openai>=2.9.0",
    "python-dotenv>=1.2.1",
    "streamlit>=1.52.1",
    "requests>=2.31.0",
    "google-genai>=1.55.0",
]
